move: dev files to detail folder
