+++ src/events/detail/default_events.py
